lambda/quantum_task_logger/requirements.txt
# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
# SPDX-License-Identifier: MIT-0

amazon-braket-sdk
aws-lambda-powertools